:zap: ::  유틸리티 모듈에 로그인 타입 열거형 명시
