+++ Projects/Feature/SignInFeature/Demo/Sources/Views/LoginButtonView.swift
@@ -11,48 +11,6 @@ import SnapKit
 import Then
 
 
-public enum LoginType {
-    
-    case kakao
-    case naver
-    case google
-    case apple
-    
-    var image: UIImage {
-        switch self {
-            
-        case .kakao:
-            return UIImage(systemName: "home")!
-            
-        case .naver:
-            return UIImage(systemName: "home")!
-            
-        case .google:
-            return UIImage(systemName: "home")!
-            
-        case .apple:
-            return UIImage(systemName: "home")!
-        }
-    }
-    
-    var text: String {
-        switch self {
-            
-        case .kakao:
-            return "카카오 로그인"
-            
-        case .naver:
-            return "네이버 로그인"
-            
-        case .google:
-            return "Google 로그인"
-            
-        case .apple:
-            return "Apple 로그인"
-        }
-    }
-}
-
 class LoginButtonView: UIView {
 
     /*
